--- TagsCloudContainer/BlackListFilterSettings.cs
@@ -1,15 +1,0 @@
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-
-namespace TagsCloudContainer
-{
-    public class BlackListFilterSettings
-    {
-        public HashSet<string> Blacklist;
-
-        public BlackListFilterSettings(IUI ui)
-        {
-            Blacklist = new HashSet<string>(new TxtWordsReader().ReadWords(ui.BlacklistPath));
-        }
-    }
-}
