+++ TagsCloudContainer/Filters/IWordsFilter.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace TagsCloudContainer.Filters
+{
+    public interface IWordsFilter
+    {
+        List<string> Filter(List<string> words);
+    }
+}
