+++ TagsCloudContainer/Formatters/IWordsFormatter.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace TagsCloudContainer.Formatters
+{
+    public interface IWordsFormatter
+    {
+        List<string> Format(List<string> words);
+    }
+}
