--- TagsCloudContainer/IWordsFilter.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace TagsCloudContainer
-{
-    public interface IWordsFilter
-    {
-        List<string> Filter(List<string> words);
-    }
-}
