--- TagsCloudContainer/IWordsFormatter.cs
@@ -1,7 +1,0 @@
-namespace TagsCloudContainer
-{
-    public interface IWordsFormatter
-    {
-        string Format(string word);
-    }
-}
