--- TagsCloudContainer/IWordsReader.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace TagsCloudContainer
-{
-    public interface IWordsReader
-    {
-        List<string> ReadWords(string path);
-    }
-}
