+++ TagsCloudContainer/Layouting/ITagsCloudLayouter.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+
+namespace TagsCloudContainer.Layouting
+{
+    public interface ITagsCloudLayouter
+    {
+        ITagsCloud TagsCloud { get; }
+        Rectangle PutNextRectangle(Size rectangleSize);
+    }
+}
