+++ TagsCloudContainer/Reading/IWordsReader.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace TagsCloudContainer.Reading
+{
+    public interface IWordsReader
+    {
+        List<string> ReadWords(string path);
+    }
+}
