+++ TagsCloudContainer/TagsCloudContainer.csproj
@@ -47,6 +47,10 @@
     </Reference>
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="mscorlib" />
+    <Reference Include="NHunspell, Version=1.2.5554.16953, Culture=neutral, PublicKeyToken=1ac793ea843b4366">
+      <HintPath>..\packages\NHunspell.1.2.5554.16953\lib\net\NHunspell.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="nunit.framework, Version=3.11.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb">
       <HintPath>..\packages\NUnit.3.11.0\lib\net45\nunit.framework.dll</HintPath>
       <Private>True</Private>
@@ -71,36 +75,40 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CircularCloudLayouter.cs" />
-    <Compile Include="CLI.cs" />
-    <Compile Include="BlackListFilterSettings.cs" />
+    <Compile Include="Filters\BlackListFilterSettings.cs" />
+    <Compile Include="Filters\BlacklistWordsFilter.cs" />
+    <Compile Include="Filters\IWordsFilter.cs" />
+    <Compile Include="Formatters\IWordsFormatter.cs" />
+    <Compile Include="Formatters\ToInitFormFormatter.cs" />
+    <Compile Include="Formatters\ToLowerCaseFormatter.cs" />
     <Compile Include="ITagsCloud.cs" />
-    <Compile Include="IUI.cs" />
-    <Compile Include="IWordsReader.cs" />
-    <Compile Include="IWordsFilter.cs" />
-    <Compile Include="IWordsFormatter.cs" />
-    <Compile Include="FrequencyWordsWeighter.cs" />
-    <Compile Include="IWordsWeighter.cs" />
+    <Compile Include="Layouting\CircularCloudLayouter.cs" />
+    <Compile Include="Layouting\ITagsCloudLayouter.cs" />
+    <Compile Include="Layouting\TagsCloudLayouterSettings.cs" />
+    <Compile Include="Reading\IWordsReader.cs" />
+    <Compile Include="Reading\TxtWordsReader.cs" />
     <Compile Include="TagsCloud.cs" />
     <Compile Include="TagsCloudGeneratorSettings.cs" />
     <Compile Include="TagsCloudContainerApplication.cs" />
-    <Compile Include="TagsCloudLayouterSettings.cs" />
     <Compile Include="Tests\TagsCloudGeneratorTests.cs" />
-    <Compile Include="TxtWordsReader.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TagsCloudGenerator.cs" />
-    <Compile Include="BlacklistWordsFilter.cs" />
-    <Compile Include="ToLowerCaseFormatter.cs" />
+    <Compile Include="UI\CLI.cs" />
+    <Compile Include="UI\IUI.cs" />
     <Compile Include="Visualisation\ArchimedeanSpiral.cs" />
     <Compile Include="Visualisation\ImageSettings.cs" />
-    <Compile Include="Visualisation\ITagsCloudLayouter.cs" />
     <Compile Include="Visualisation\ITagsCloudRenderer.cs" />
     <Compile Include="Visualisation\PNGTagsCloudRenderer.cs" />
     <Compile Include="Visualisation\TagsCloudWord.cs" />
+    <Compile Include="Weighting\FrequencyWordsWeighter.cs" />
+    <Compile Include="Weighting\IWordsWeighter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="nhunspell" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
