--- TagsCloudContainer/ToLowerCaseFormatter.cs
@@ -1,10 +1,0 @@
-namespace TagsCloudContainer
-{
-    public class ToLowerCaseFormatter : IWordsFormatter
-    {
-        public string Format(string word)
-        {
-            return word.ToLower();
-        }
-    }
-}
