+++ TagsCloudContainer/UI/IUI.cs
@@ -1,0 +1,15 @@
+using System.Drawing;
+
+namespace TagsCloudContainer.UI
+{
+    public interface IUI
+    {
+        string InputPath { get; }
+        string OutputPath { get; }
+        string BlacklistPath { get; }
+        Point TagsCloudCenter { get; }
+        Size LetterSize { get; }
+        Size ImageSize { get; }
+        Color TextColor { get; }
+    }
+}
