--- TagsCloudContainer/Visualisation/ITagsCloudLayouter.cs
@@ -1,10 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudContainer.Visualisation
-{
-    public interface ITagsCloudLayouter
-    {
-        ITagsCloud TagsCloud { get; }
-        Rectangle PutNextRectangle(Size rectangleSize);
-    }
-}
