+++ TagsCloudContainer/Weighting/IWordsWeighter.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Drawing;
+
+namespace TagsCloudContainer.Weighting
+{
+    public interface IWordsWeighter
+    {
+        Dictionary<string, Size> GetWordsSizes(List<string> words, Size minLetterSize);
+    }
+}
