+++ TagsCloudContainer/nhunspell/ru/README.txt
@@ -1,0 +1,8 @@
+
+	Russian Hunspell dictionary
+
+Автор:
+	Александр Клюквин (https://sites.google.com/site/dictru/)
+Сайт проекта:
+	http://code.google.com/p/hunspell-ru/
+
