+++ TagsCloudContainer/nhunspell/ru/settings.yml
@@ -1,0 +1,3 @@
+---
+ignore_case: false
+
