+++ TagsCloudContainer/packages.config
@@ -3,6 +3,7 @@
   <package id="Autofac" version="4.9.0-beta1" targetFramework="net461" />
   <package id="CommandLineParser" version="2.3.0" targetFramework="net461" />
   <package id="FluentAssertions" version="5.5.3" targetFramework="net461" />
+  <package id="NHunspell" version="1.2.5554.16953" targetFramework="net461" />
   <package id="NUnit" version="3.11.0" targetFramework="net461" />
   <package id="System.Collections" version="4.0.11-rc2-24027" targetFramework="net461" />
   <package id="System.Console" version="4.0.0-rc2-24027" targetFramework="net461" />
