+++ hunspell_ru_ru/expected/hunspell_ru_ru.out
@@ -1,0 +1,78 @@
+CREATE EXTENSION hunspell_ru_ru;
+CREATE TABLE table1(name varchar);
+INSERT INTO table1 VALUES ('земля'), ('землей'), ('землями'), ('земли'),
+						('туши'), ('тушь'), ('туша'), ('тушат'),
+						('тушью');
+SELECT d.* FROM table1 AS t, LATERAL ts_debug('russian_hunspell', t.name) AS d;
+ alias |    description    |  token  |          dictionaries           |    dictionary    |        lexemes         
+-------+-------------------+---------+---------------------------------+------------------+------------------------
+ word  | Word, all letters | земля   | {russian_hunspell,russian_stem} | russian_hunspell | {земля}
+ word  | Word, all letters | землей  | {russian_hunspell,russian_stem} | russian_hunspell | {земля}
+ word  | Word, all letters | землями | {russian_hunspell,russian_stem} | russian_hunspell | {земля}
+ word  | Word, all letters | земли   | {russian_hunspell,russian_stem} | russian_hunspell | {земля}
+ word  | Word, all letters | туши    | {russian_hunspell,russian_stem} | russian_hunspell | {туша,тушь,тушить,туш}
+ word  | Word, all letters | тушь    | {russian_hunspell,russian_stem} | russian_hunspell | {тушь}
+ word  | Word, all letters | туша    | {russian_hunspell,russian_stem} | russian_hunspell | {туша,тушить,туш}
+ word  | Word, all letters | тушат   | {russian_hunspell,russian_stem} | russian_hunspell | {тушить}
+ word  | Word, all letters | тушью   | {russian_hunspell,russian_stem} | russian_hunspell | {тушь}
+(9 rows)
+
+CREATE INDEX name_idx ON table1 USING GIN (to_tsvector('russian_hunspell', "name"));
+SELECT * FROM table1 WHERE to_tsvector('russian_hunspell', name)
+	@@ to_tsquery('russian_hunspell', 'землей');
+  name   
+---------
+ земля
+ землей
+ землями
+ земли
+(4 rows)
+
+SELECT * FROM table1 WHERE to_tsvector('russian_hunspell', name)
+	@@ to_tsquery('russian_hunspell', 'тушь');
+ name  
+-------
+ туши
+ тушь
+ тушью
+(3 rows)
+
+SELECT * FROM table1 WHERE to_tsvector('russian_hunspell', name)
+	@@ to_tsquery('russian_hunspell', 'туша');
+ name  
+-------
+ туши
+ туша
+ тушат
+(3 rows)
+
+DROP INDEX name_idx;
+CREATE INDEX name_idx ON table1 USING GIST (to_tsvector('russian_hunspell', "name"));
+SELECT * FROM table1 WHERE to_tsvector('russian_hunspell', name)
+	@@ to_tsquery('russian_hunspell', 'землей');
+  name   
+---------
+ земля
+ землей
+ землями
+ земли
+(4 rows)
+
+SELECT * FROM table1 WHERE to_tsvector('russian_hunspell', name)
+	@@ to_tsquery('russian_hunspell', 'тушь');
+ name  
+-------
+ туши
+ тушь
+ тушью
+(3 rows)
+
+SELECT * FROM table1 WHERE to_tsvector('russian_hunspell', name)
+	@@ to_tsquery('russian_hunspell', 'туша');
+ name  
+-------
+ туши
+ туша
+ тушат
+(3 rows)
+
